Move use pending under shared
