+++ app/src/editor/canvas/components/CanvasController/useCanvasCreateCallback.jsx
@@ -2,14 +2,14 @@ import { useContext, useCallback } from 'react'
 
 import useIsMountedRef from '$shared/utils/useIsMountedRef'
 import { Context as RouterContext } from '$shared/components/RouterContextProvider'
-import usePending from '$editor/shared/hooks/usePending'
+import usePending from '$shared/hooks/usePending'
 
 import links from '../../../../links'
 import * as services from '../../services'
 
 export default function useCanvasCreateCallback() {
     const { history } = useContext(RouterContext)
-    const { isPending, wrap } = usePending('CREATE')
+    const { isPending, wrap } = usePending('canvas.CREATE')
     const isMountedRef = useIsMountedRef()
 
     return useCallback(async ({ replace = true } = {}) => {
