+++ app/src/editor/canvas/components/CanvasController/useCanvasDuplicateCallback.jsx
@@ -2,14 +2,14 @@ import { useContext, useCallback } from 'react'
 
 import useIsMountedRef from '$shared/utils/useIsMountedRef'
 import { Context as RouterContext } from '$shared/components/RouterContextProvider'
-import usePending from '$editor/shared/hooks/usePending'
+import usePending from '$shared/hooks/usePending'
 
 import links from '../../../../links'
 import * as services from '../../services'
 
 export default function useCanvasDuplicateCallback() {
     const { history } = useContext(RouterContext)
-    const { isPending, wrap } = usePending('DUPLICATE')
+    const { isPending, wrap } = usePending('canvas.DUPLICATE')
     const isMountedRef = useIsMountedRef()
 
     return useCallback(async (canvas) => {
