+++ app/src/editor/canvas/components/CanvasController/useCanvasRemoveCallback.jsx
@@ -2,14 +2,14 @@ import { useContext, useCallback } from 'react'
 
 import useIsMountedRef from '$shared/utils/useIsMountedRef'
 import { Context as RouterContext } from '$shared/components/RouterContextProvider'
-import usePending from '$editor/shared/hooks/usePending'
+import usePending from '$shared/hooks/usePending'
 
 import links from '../../../../links'
 import * as services from '../../services'
 
 export default function useCanvasRemoveCallback() {
     const { history } = useContext(RouterContext)
-    const { isPending, wrap } = usePending('REMOVE')
+    const { isPending, wrap } = usePending('canvas.REMOVE')
     const isMountedRef = useIsMountedRef()
 
     return useCallback(async ({ id }) => {
